signup.php now has complete error handling.
error.php has a class to manage a "current error" object. This object
should be created in a catch block of an exception, when rudimentary
form validation fails, or other such server-side error conditions. Then
the current page should be returned to and Error::putError() will print
the error to the page.

formInput.php provides functions for rudimentary validation that forms
have been filled in, and a way to correctly get values from form fields
while dealing with whether isset() and empty() is called correctly
without producing warnings or crashing.

This begins my push to remove separate "process" pages in favor of using
same-page form targets, so that error states can be displayed on the
form without requiring a redirect.