Events are now created by Actions. There will no longer be insertEvent calls scattered through db.php in disarray.
This means that normal database interaction events should be managed by the TarsException and Action classes for us, leading to less Event logging management.

This fixes a bug where special Session-related Events could get a NULL creatorID when it's clear that a User is associated with them.

This updates the Actions $action_map to support event-log settings. For now, all Actions that modify the database in some way have 'eventLog' => 'always' (EventTypes.level = 'info' or above),
while all others will only be logged if a TarsException is created about it. This can now easily be changed so that all the lookup events (EventTypes.level = 'debug') are also logged in the future.