Made it so editTermUI uses the object system for Professors, simplifies code
Also, use professor emails here (for now) so that we can actually
associate the entered name with an actual account.

Session time/place for lectures/labs/workshops without a currently set
lecture or lab should now appear with default values. In theory when
these are saved, a Session database object would be created if and only
if these values are not kept default.