Implementing pagination into database. Database::executeGetPage($sql, $args, $pgArgs) will fetch a page of Objects for an action.
The SQL will be modified to include an ORDER BY and LIMIT statement in that function, based on the $pgArgs provided.

The JavaScript now has a doPaginatedAction() version of doAction() that prepares a page for paginated results using this system.

While unused by the JavaScript at this time, ORDER BY is supported by the executeGetPage() code

A working student/search.php, student/search.js, and actions.php:'fetchPositions' is implemented

This also changes the behavior of some internal code to allow Students to apply to Positions after they Cancelled a previous Application (not Withdrawn) and now the app is now aware of what state the user's latest application is in.